Replaced about+mp3 links at top of score page 1
A link to the "About" file for the score is now porvided at the
top of page 1 of each score, but there is no longer a link for playing an MP3 file. MP3 files should be provided in the
"About" file, if possible.
The text boxes for the MP3 file have been removed from the
Metadata dialog (in Assistant Composer).
Added a style to each page, to prevent the cursor changing to
an I-Beam every time it moves over text (e.g. a clef, notehead,
dynamic etc.).
Removed the writing of a "Javascript" section in the score.