Change to the use of durations in chords, changes in SvgPage.cs
Durations: the durations of score:basicChords no longer have to add up
to the score.msDuration of their containing score:chord in an SVG file.
This rather drastic change, which meant changing quite a number
of files, allows the Assistant Performer's speed option to work correctly.
In SvgPage.cs:
1. The Assistant Composer now adds the onload="onLoad()" attribute
and the <script xlink:href="../../ap/SVG.js" type="text/javascript"/>
element to the first page of each score.
2. Bug fix: When creating a score, an unnecessary exception was being
thrown if the metadata did not contian a link to a page about the score
(on my website). Now, a MessageBox reminder is presented instead.