Song 6 Sketch completely reorganised
Known Bug: rests are being created in Clytemnestra.cs, but
they are not currently being displayed.
Clytemnestra.cs and Song6SketchAlgorithm.cs have been
completely rewritten to allow for birds and winds voices to
be added fairly easily later.
The score currently looks as it did before this revision,
except that the rests between verses now consist of just one
bar (8 sec.)
There have been a few changes to other files:
e.g. Lyrics can now be set in MidiChordDefs, and are then
automatically added to ChordSymbol.DrawObjects when
the ChordSymbols are constructed.