Reverted to having no winds aligned for the moment.
Apart, that is, from those to which Clytamnestra
was aligned (at barlines).
Alto and soprano winds are going to be constructed
at a krystal level higher than the tenor and bass, and
_they_ will probably be aligned as well.
So will the four channels of birds.
But I need to compose alignments _as_a_whole_
when all the material exists (birds and winds).
The function MidiDefSequence.AlignChordOrRest()
has at least been tested, and can be used in future.