Changed the krystals used. (simplification)
Instead of truncating xk3(7.7.1)-3.krys to create the winds,
I have now created a new krystal, with corresponding
input krystals, which contains the same results.
The score has not changed, but the krystal is simpler,
and its density inputs can be used to create more
krystals in the same series at higher levels.