Completed MidiPhrase.SetContour(...)
This is a very powerful contouring function which
re-orders the LocalMidiDurationDefs in (part of) a
MidiPhrase. I've debugged it, and it seems to be
working, except that I have not tried including
rests in the permuted part of the MidiPhrase.
There is extensive documentation above the
setContour function itself.
Also: moved the function Winds.SetLyricsToIndex()
inside MidiPhrase.