First pass for Furies1Finale
The Furies1Finale still needs moulding into shape, but
the basic structure is there. TODO: organize the durations,
and pitches (contours?).
PitchwheelDeviation has been done but probably needs
adjusting again when all FuryFinales are complete.
When all the FuryFinales are complete, first adjust the
alignments and velocities (Velocities should be reduced
when Clytemnestra is speaking.), and then compose the
controller (R2M).