Bug Fix: Patch indices were being written too often in MidiChordDefs.
A Patch index will now only be written as an attribute of the main
MidiChordDef if MidiChordDef.BasicMidiChordDef[0].PatchIndex is
not set.
(Whether a Patch change message is actually contructed and sent
by the AssistantPerformer should depend on the current patch in
the voice.)
The same is also true for Bank indices.