Setting MsDuration of a UniqueMidiChordDef sets BasicMidiChordDefs
The durations of UniqueMidiChordDefs are consistently equal to
the sums of the durations of the contained BasicMidiChordDefs.
A UniqueMidiChordDef's msDuration is no longer written to SVG
files, but is calculated from the BasicMidiChordDefs' msDurations
by the Assistant Performer.
All functions that change the msDuration of a
UniqueMidiChordDef do so by setting its MsDuration property,
which changes the durations (and possibly number) of the
contained BasicMidiChordDefs