Added stub code for Study 4
The object of this exercise is to write a new piece
using an improved version of the VoiceDef class.
1. Added a copy of VoiceDef.cs to the SongSix
folder, and renamed it SongSixVoiceDef.cs.
Refactored all the Song Six code to use this new
SongSixVoiceDef class.
2. The original VoiceDef class still exists, and one
of its constructors is still used by Study 3 sketch.
But this VoiceDef is going to change (probably
quite drasically) to take account of the Study 4
algorithm. I want to make the process of
composing easier by defining positions in the
score symbolicly (i.e. names instead of numbers).
3. The file Study4Algorithm.cs exists, but is simply
a cannibalised copy of the original
SongSixAlgorithm.cs which has been crudely
edited so that it compiles.
4. The Study4.mkss file exists, but needs changing.