Partially reverted to previous commit.
The track initialization values should be set in
their own dialog (like the PerformerOptionsForm)
and not on the AssistantComposerMainForm.
Also, the trackInit and performerOptions elements
and their attributes should only be written/read
if they are not the default values.
This commit does not yet implement the dialog
for setting track initialization values.