Added a stub PolyPerformerOptionsForm
Renamed the existing PerformerOptionsForm to
MonoPerformerOptionsForm, and added a new
PolyPerformanceOptionsForm whose public
interface is the same as the mono version, but
whose public functions are not yet implemented.
Added a ComboBox to the main Assistant
Composer form to choose between 'none',
'mono' and 'poly' event handlers.
The AssistantPerformer's event handler will be
selected according to whether the .mkss file
contains a monoPerformanceOptions element,
a polyPerformanceOptions element, or neither.
The AssistantComposerMainForm itself should
be working for all defined types of event handler.