Unified the PerformersOptionsForm.
There are no longer two Forms for mono or poly
performer options.
If/when key options are needed for the keyboard1
event handler, they will be added as optional
controls in the existing PerformersOptionsForm.
Removed the old noteOnPitch and noteOnVelocity
track options, since these were not used, and will
probably be superceded by options for individual
keys.
Recompiled a couple of scores and their .mkss
files. Working.