Revised Preferences, recompiled all scores
The preferences dialog can now only set
the preferred midi input and output devices,
and these are the only values stored in the
preferences file. However, the standard paths
to the environment files and folders are
displayed for reference in the preferences
dialog. These strings vary, depending on
whether Moritz is running on my AUDIOPC.
or on some other computer (e.g. my laptop
I7ACERLAPTOP). The strings are set in
Moritz.Statics.cs.
Also did quite a lot of relevant refactoring.