Assistant Composer for "prepared piano" begins.
Scores will contain both input and output staves.
The input staves tell the performer which keys to
press. Output staves contain midi information
which will be sent to the output device under
control of the info provided by the performer.
Initial development is by realising a new score:
Study 3 sketch 2, which adds a new input staff
to Study 3 sketch 1. The input staff will have a
1-note chord for each sequence to be triggered
in the output staves.
*************************************************
This commit compiles, and all scores prior to
Study 3 sketch 2 have been re-compiled.
The AssistantComposerMainForm has been
reprogrammed, and loads the Study 3 sketch 2
setttings correctly, but the score will not yet
compile because the voice.UniqueMidiDurationDefs
list needs to be changed to a voice.NoteObjectDefs
list which will accept InputChordDef and
InputRestDef objects (or something to that effect).
See Notator.cs.AddSymbolsToSystems(...).
***************************************************