SVG now uses "class" instead of "id" attributes.
Currently finalizing the new version of the
SVGExtensions file format.
The Assistant Performer has also been updated
for this version of the file format (though input
staves have not been created or tested anywhere),
and the current scores all play back correctly.
(Known bug in the AP: performances that start
after the beginning are not working correctly.)
The change to using "class" instead of "id" seems
logical. The "id" attribute is now only written to
"SvgSystem" and "Staff" objects (as a
convenience when debugging SVG files).
Also fixed some bugs.