Intermediate Commit (Study 3 sketch 2 begins)
The Study 3 sketch 1 algorithm has been
extended to include an Input Staff, thus creating
the Study 3 sketch 2 algorithm.
A large number of changes have also been made
to the other code so that it takes account of Input
staves, voices and chords.
The score currently compiles, but there are
some problems in the graphics (stems are in the
wrong direction, cautionary chords are not yet
correct...).
Also, I have not yet begun to program
performance options or the content of the score:inputChord SVG element.