PerformanceControlDef defined and first used.
The performance control options are now set in
composition algorithms, and the
PerformersOptionsForm in the AssistantComposer
is no longer used.
The PerformersOptionsForm is therefore
completely redundant, and should be removed.
Do this, and clean up, before testing the new
options further.
Study 3 sketch 2 contains a first implementation
of the new class, but this needs further testing and
verification here in Moritz before putting the
options into practice in the AssistantPerformer.
The values in PerformanceControlDef should
cascade through the Score, InputVoice and
InputChord classes.
The global performance control options are now
written to the first SVG page in the score, no
longer to the .mkss file.