PerformanceControlOptions revision
The PerformersOptionsForm in the
AssistantComposer is still completely redundant,
and should be removed. Do this soon!

PerformanceControlOption objects no longer
"cascade". They can now be an attribute of either
OutputVoice or OutputChord, and the
OutputChord settings update the current
OutputVoice settings as they occur.
The Study3Sketch2Algorithm now uses this new
paradigm. (There were problems with the old one,
relating to the indeterminacy of the order of
noteOns in a chord.)
MasterVolumes are treated similarly.