InputControls now override completely
InputControls element are now always written
completely. These options are then valid
in the voice until further notice. This way it is
much easier to keep track of which options are
currently in force.
For test purposes in Study 3 sketch 2, added
pitchwheel->volume controls to bar 3, and
limited fade controls to bar 4.