Excluded the main AssistantPerformer Form
The C# version of the assistant performer is no
longer supported. However, code in the
Moritz.Performer project/namespace is used to
play sounds in palettes. This is currently not
working, and is being debugged.
Reprogrammed MoritzForm1 so that it can no
longer open an AssistantPerformer Form.
The AssistantPerformer Form is no longer being
programmed, but it is still being tracked by Git
(for archive purposes).