OutputVoice.VoiceID property added.
This property is null if there are no InputVoices
in a score, otherwise it is the index of the
OutputVoice's VoiceDef in the bar definitions
originally constructed by an algorithm.
It is necessary to have this ID separate from
the midi channel index, because midi channels
are always in ascending order (top-down), even
when the voices are printed in a different order.
InputNotes must be able to reference the
proper voice, wherever it is in the system.
Major revision of the comment to
CompositionAlgorithm.DoAlgorithm().