alignmentX attributes removed.
alignmentX attributes, which used to be
written to InputChordSymbol, RestSymbol
and OutputChordSymbol, are no longer
written anywhere in score files.
Instead, InputChordSymbol writes its
msDuration to a new inputChord:msDuration
attribute. (The msDuration is known, and has
already been used to find the chord's
duration class.)
The Assistant Performer will now use the
common msPosition to link InputChords
to Seqs. This is much safer, and leads to
more elegant code.