Moved InputControls inside InputChordDef
This means that the output voices in a score
are identical, whether or not there are any
input voices in the score.
This leads to simpler, more elegant code.
(Less need for consistency checking.)
Also revised and updated the new
svgScoreExtensions document, and did
some refactoring.
The "score:"prefix is now only written for
attributes and elements in the score when
neccessary.
e