Completed  the PaletteChordForm
The final step was to add buttons for perfoming
the same audio files as in the containing
PaletteForm.PaletteControl.AudioSampleButtons.
The new buttons are directly linked to the
AudioSampleButtons, simply passing on the
arriving mouse events. The AudioSampleButtons
contain all the functionality.
Also a couple of small bug fixes in the
PaletteChordForm.