Restructuring Saving: OrnamentSettingsForm
beginning to implement a new saving
strategy for .mkss files.
The OrnamentSettingsForm is complete, but
otherwise the following restructuring is still
incomplete.
Each of the Assistant Composer's subsidiary
forms is going to be given an OkayToSave and
RevertToSaved button, implementing a new
.mkss saving strategy.
Each of these forms will manage its own 'saved
status', setting its title text accordingly.
Before the AC saves, it is going to check that
none of the forms' titles ends with a "*". If a
form's title does end with a "*", then the form
has to be reviewed, and either its OkayToSave or
RevertToSaved button has to be clicked to
remove the "*".