Continued work on AssistantComposerMainForm
There are now separate buttons implementing
IRevertable for the notation, krystals and
palettes panels. These currently work properly
with each other and with the form as a whole.
Revised the layout of the other buttons too:
RevertCompletely, ShowCheckedForms and
ShowUncheckedForms have not yet been
implemented.
The AssistedComposerMainForm_Activated()
event handler exists, but does not yet do
anything.