ReviewableFormFuntions as singleton
There is now a single object of this type. It is
constructed as part of the AssistantComposerForm
and passed as an argument to the subsidiary forms'
constructors.
Added some new accessor functions to the
ReviewableFormFuntions interface, so that its
internal lists (which are no longer static) are kept
private.
Bug Fixes: The paletteForm and ornamentForm
revert functons now work correctly.