Scores can now contain invisible output staves.
If there is at least one input voice in a score, all
the output voice midi information is now written
into scores. That output information may however
be invisible, i.e. not associated with any graphic
info.
Each invisible voice is contained in an invisible
staff in the system. Invisible staves are written
above the visible staves in svgSystems. The midi
info they contain is organised identically.
TODO: update the AssistantComposerForm, and
adjust the help texts.