Vertical system justification improvement.
The distance between the bottom staffline in one
system and the top staffline in the next, is now a
multiple of pageFormat.Gap. (The distance
between stafflines inside systems is already a
multiple of pageFormat.gap.) This means that
stafflines are now always vertically aligned to
screen pixels, and therefore look better.