An InputVoice can now have a midi channel
In future, multiple input devices may be possible,
so there is no longer any upper limit to the number
of input voices.
InputVoice midi channel attributes are written only
if an InputVoiceDef.MidiChannel is explicitly set
(by an algorithm), and then only to the input voices
in the first system in a score.