Completed changes for Inkscape
Updated the info set in Inkscape's metadata
for each page. This now includes the link to
the score's 'about' page. This link is no longer
written on page 1 of each score. (It is, however,
now written on the AP's main dialog when a
score is selected.)
Five layers are now created (bottom to top)
layer1, "boundary": the line round the page.
layer2, "moritz": all moritz' score output
layer3, "score annotations": Currently empty,
but will be used for annotations that are
considered part of the composition (slurs,
performance instructions etc.)
layer4, "user annotations": Currently empty,
but will be used by performers, if they need
to add any annotations of their own.
layer5, "transparent, clickable surface":
this contains a transparent rectangle covering
the page. This is used while setting the start-
and end-markers in the Assistant Performer.
The use of layer5 enables the AP to improve
the behaviour of the set start marker and set
end marker tools: these no longer revert to an
I-beam when passing over text...
Recompiled all scores.