Reorganised staff connectors etc. (see below)
Staff connectors are now created in a group having
no class, but an id: sys<systemNumber>connectors.
The rightmost barlines and connectors in voices and
systems are now drawn to take account of staffline
thickness (i.e. without notches).
To be future-proof, the SystemMetrics boundary
(i.e. its Top, Left, Bottom and Right values) now
takes account of staffName and barnumber metrics.
(The SystemMetrics boundary is not currently *used*.)