Tidied up class and id attributes
Only the objects that need a class attribute in the
Assistant Performer are now actually given one.
These classes are:
        system,
        outputStaff,
        inputStaff,
        stafflines,
        outputVoice,
        inputVoice,
        outputChord,
        inputChord,
        rest.
Many other objects are given an id, consting of
their type concatenated with a unique number.
These ids are not used by the Assistant Performer,
but are useful as comments when editing in
Inkscape.