InputOptions improvements
Removed unused options completely.
All inputOption enums now have both
'undefined' and 'off' members.
'undefined' is the default value, not written
to the score, meaning "use the current value
of this option".
'off' is written to the score, and means
"this option is now turned off, use the values
in the score"
Also: the NoteOnVel option now requires a
'minVelocity' value to be defined and written
to the score (if it is not being turned 'off').