Synchronous ccSettings check added
1. Scores may not contain synchronous ccSettings,
so I've added a check for this to the existing
checks in ComposableScore.CheckBars(...).
2. Moved the code for setting voice channels
and master volumes from CompositionAlgorithm
to ComposableScore. This means that these values
are automatically moved to the correct places in
scores. Algorithm code does not need to set them.
3. Moved the GetInputChordDefsInBar(bar) code
from Study3Sketch2Algorthm.cs into the base class
CompositionAlgorithm. More useful there in future.