@font-face definitions now included in .svgs
Moritz now writes the @font-face definitions
of CLicht, Arial, Open Sans, and
Open Sans Condensed in a <style> element
inside each .svg file's <defs> section.
This means that fonts should load correctly
when the svgs are added as <object>s to
HTML files. This allows scores to be displayed
in the documentation, outside the Assistant
Performer.