Patch 0 set in all chords in Pianola Music
In other score algorithms, midiChordDef.Patch
 is always set. This is done here too, for consistency,
even though the patch does not change, and it is
therefore not necessary to reset the patch when
restarting a performance.
If the midiChordDef.Patch is not set on every chord,
and the performance starts from somewhere in the
middle after performing another piece, the channels
will all have the wrong patch.