Moved the score construction site
Scores are now created in the
assistantPeformerTestSite/scores directory (where
their .mkss files are also kept).
When complete, the scores will be copied to the
public assistantPeformer/scores directory, from
which they can be published.
This simplifies things by keeping the score of
Tombeau 1 private while it is being composed.
(The TestSite version of the AssistantPerformer can
include Tombeau 1 in its list of scores.)