Defined Seq.Concat(seq2)
Concat(seq2) concatenates a clone of seq2
to the caller (and returns a pointer to the
caller. This is analogous to string.Concat(str2) ).
The Seq(seq1, seq2) constructor now makes
a clone of seq1, then calls seq1Clone.Concat(seq2)
before setting the new Seq's fields to those of
seq1Clone.