Study 1 working again (changed dynamcs)
The dynamic symbols are now related
logarithmically to the midiVelocity.
This corresponds to the way gain is usually
rendered by MIDI synthesizers.
In Study 1, the dynamic symbols are now
as they were in the original score, while the
velocities have changed.
This change in the transcription of velocities
is going to apply to all scores as they are
recompiled.