Study 1, Pianola Music and Tombeau 1 working
All the above scores display and play
correctly in the Assistant Performer. Including
the greying out of disabled voices.
This commit includes further corrections that
needed to be made as a result of the change to
local time coordinates in Trks, Seqs and their
containers. The names of all msPosition
attributes now say what their frame of reference
is. Bug Fixes:
1: Flags are now drawn.
2: System widths are now calculated correctly
even when the final chord is split onto the
next system.