Velocity->dynamic correspondence now final.
Moritz now uses dynamic symbols in the range
[pppp->fff], each dynamic symbol corresponding
to an equal sized range of velocities. Study 1's
dynamics are as they were in the original score,
but the velocities have changed slightly.
In the other scores, the velocities have not changed,
but their transcriptions as dynamics have.
Recompiled all scores except Song Six.