Notehead colour can now denote velocity.
This is a large commit, affecting many files.
The code has been tested, and works, in all
the test scores.
There is a new pop-up menu in the Assistant
Composer's main form for setting the output
chord symbol type. This pop-up has two values:
"standard black" and "coloured velocities". This
value is saved in .mkss files. (Input chord symbols
have no midi velocity, so their noteheads and
extenders are never coloured anything but black.)
If "coloured velocities" is selected, (output)
noteheads and extenders are coloured according
to the note velocities. The densities and velocities of
BasicChordDefs depend on ornament definitions,
so are independent of the notated values.
Added functions for setting the velocities of notes
in individual MidiChordDefs.
Updated the main comment in Tombeau1Algorithm.cs