Trks and InputVoiceDefs have Container attribute
The Container is an IVoiceDefContainer interface.
This enables Trks to be manipulated even when
they are contained in Seqs. They can not be
changed when inside Blocks.
Trk.Permute(...) has been revised and tested.
Tombeau 1 has been used to test all the Trk
functions to see that they work correctly.
All the other test scores compile correctly too.