Envelope.cs enhancements.
An envelope is a list of integral values that
are greater than or equal to 0, and less than
or equal to the envelope's UpperBound.
The list's Count must be greater than 0, but
is otherwise unlimited.
UpperBound must be >= 0.
MidiChordDef Sliders use envelopes having
UpperBound==127.
The list's values can repeat. Not all values in
range have to be present.
Values that are set to 0 in envelope.Original
are set to UpperBound in envelope.Inversion.
Values that are set to UpperBound in
envelope.Original are set to 0 in
envelope.Inversion.