Added MidiChordDef.SetOrnament(...)
This function creats countArg BasicMidiChordDefs,
having equal msDuration, in the MidiChordDef
without changing its msDuration.
Each BasicMidiChordDef has a single pitch having
velocity 127.
The pitch is NotatedMidiPitches[0] + an interval
found in the pitchHierarchy.
The index of the interval is found using the envelope.
Added some test code to Init.cs.