MoritzStatics.GetAscendingPitches(...) changed
The current version returns a list of pitch numbers
in range [0..127] in ascending order.
The first pitch in the returned list is always rootPitch.
The other returned pitches are transposed to be in
ascending order by adding 12 as necessary.
The maximum number of pitches returned is either
nPitches or the number of pitches beginning with and
following the rootPitch in the absolutePitchHierarchy,
whichever is smaller.
        /// The number of returned pitches can also be smaller than nPitches because pitches that would be higher
        /// than 127 are simply not added to the returned list.