UpsideDown() functions in MidiChordDef and BasicMidiChordDef
MidiChordDef.UpsideDown() returns a new MidiChordDef
containing a single BasicMidiChordDef having the same
NotatedMidiPitches.
The original base pitch is preserved, but the top-bottom
order of the prime intervals is reversed.
Velocities remain in the same order, bottom to top.
To add upside-down BasicMidiChordDefs, call the following
function:
BasicMidiChordDef.UpsideDown(): this returns a new
BasicMidiChordDef having the original base pitch,
but in which the top-bottom order of the prime intervals
is reversed. Velocities remain in the same order, bottom to top.