Refactored to Block, Seq and Trk.SetPitchWheelSliders(...)
This is now a function that takes an Envelope argument
in Block, Seq and Trk.
There is also an overloaded version in Trk that takes a
Dictionary<int, int> pitchWheelValuesPerMsPosition
argument. This overload is called by the other three
functions to set pitchWheel values in MidiChordDefs
at the Dictionary positions.
Updated Tombeau1 main comment.