MidiChordDef.Invert(nPitchesToShiftArg)
In each chord in the MidiChordDef:
1. nPitchesToShift is set to nPitchesToShiftArg
    mod nPitches in the chord.
2. nPitchesToShift pitches are shifted up one octave.
3. The pitch list is resorted to be in ascending order.
Notes:
a) Velocities do not change. They remain in the same
order as they were before this function was called.
b) The pitches thereby remain part of the Gamut, if
there is one.
Deleted old (commented out) Boulez addition from
MidiChordDef.cs
Renamed MidiChordDef.Transpose(Gamut...) to
TransposeInGamut(...), and added a Trk function of
the same name, that calls thefunction on the
contained MidiChordDefs.
TestBlocks updated and running.