Clef changes can now be composed in Trks.
This need some further testing for clef changes
at other positions, but I think it is basically
correct.
Is the space to the left of clef changes correct?
What about clef changes that happen at the
beginnings and ends of Trks? (=VoiceDefs)
Also: changed the strategy in Block1TestBlock.cs
for creating Blocks and Seqs.
This makes Trk alignment more convenient.