Changed CompositionAlgorithm.InitialClef handling.
Initial Clefs are now automatically those defined in the
page format. So, when writing algorithms, one no longer
has to copy the initial clefs manually to the SetInitialClefs
function.
(Algorithms now have an InitialClefs property that is set
automatically before the algorithm is run, and the Block.SetInitialClefs function is now private.)