StartBlock and Target1Block are now classes
Each is a subclass of Block. This means that their
construction code is encapsulated. Private functions
are no longer just inside the Tombeau1Algorithm class.
Note that, contrary to the Block and MainBlock classes,
StartBlock and Target1Block are in the Tombeau1
*namespace*. This is a *pattern*.
Tombeau1's blocks can be composed independently
of each other, and of blocks composed for other
compositions.