Enhancements to SetDuration and SetVelocity functions
The following functions have been given an
optional percent argument (default value 100):
MidiChordDef.SetVelocityFromDuration(...)
MidiChordDef.SetVelocityPerAbsolutePitch(...)
Trk.SetVelocitiesFromDurations(...)
Trk.SetVelocityPerAbsolutePitch(...)
Trk.SetDurationsFromPitches(...)
In each case, the percent argument determines
the proportion of the final value for which the
function is responsible. The other component
of the final value is its existing value. If percent
is 100.0, the existing value is replaced completely.