Trk.Superimpose(trk2) function added.
This function attempts to add all the non-RestDef UniqueDefs
in trk2 to the calling Trk at the positions given by their
MsPositionReFirstIUD added to trk2.MsPositionReContainer,
whereby trk2.MsPositionReContainer is used with respect to
the calling Trk's container.
Before doing the superimposition, the calling Trk is given
leading and trailing RestDefs so that trk2's uniqueDefs can be
added at their original positions without any problem there.
These leading and trailing RestDefs are however removed
before the function returns.
The superimposed uniqueDefs will be placed at their original
positions if they fit inside a RestDef in the original Trk. If this is
not the case, they are interspersed wth the original uniqueDefs
in order of the original positions, so that the original positions
in both Trks will change.