Added FinalizeSeqs.cs
Its important, while composing the trks, to realise
that there is going to be a "second pass" through
the seqs. Each of the four channels should pursue
its own development fairly independently, but their
final alignment and integration needs to be done
vertically inside each seq to optimise the
coincidences.
Also put the GetGamuts() function in its own file.