SetVelocityPerAbsolutePitch() revised...
in Trk, MidiChordDef and BasicMidichordDef classes.
Using a velocityPerAbsolutePitch from a Gamut that
is not the one used to construct a MidiChordDef in
the first place can result in BasicMidiChordDefs
consisting of a single note having pitch==0,
velocity==0.
If a MidiChordDef containing no velocities > 0
results, both its NotatedMidiPitches and
NotatedMidiVelocities will be empty.
If this is the case when MidiChordDef.SetVelocityPerAbsolutePitch()
returns, the caller (Trk.SetVelocityPerAbsolutePitch() )
replaces the MidiChordDef by a RestDef having the
same duration.