Added the Grp class (derived from Trk)
Grps are beamed groups of UniqueDefs that
can be added to a Trk. The final MidiChordDef
in a Grp has BeamContinues set to false. All
the other MidiChordDefs have BeamContinues
set to true.
Grp must therefore override all the Trk functions
that change the MidiChordDefs' sequence or
existence.
TODO:
1. Complete the overriding of such functions.
2. All the functions that change velocities should
allow velocity 0, and behave similarly to the current
behaviour of SetVelocityPerAbsolutePitch().
3. Complete the SopranoTrks, AltoTrks and
BassTrks classes analogously to TenorTrks,
removing the SopranoTemplate,
AltoTemplate and BassTemplate classes.