This branch generates Tenor Grp palettes
A palette is a list of Grps that can be cloned
and adjusted for use in the composition.
Each of the 22 bars contains 7 Grps that all
share the same Gamut. Each Grp has as many chords as there are pitches per octave in its
gamut.
The first chord in the first Grp in each bar has
root pitch 48, since C natural is common to
all gamuts. Successive root pitches in the first
Grp are the successive pitches in the gamut's
AbsolutePitchHierarchy.
Successive Grps in the bar are transposed
(in the gamut) by gamut steps that are
successive pitches in the gamut's
AbsolutePitchHierarchy.
The Trks constructor and FinaliseSeqs() can
use any of the availabe transformation code,
including TransposeInGamut(), SetVelocity...,
SetDuration..., Permutation, TimeWarp etc.