Study 1 compiles and is written correctly
Much debugging still has to be done.
There are known bugs in compiling the Pianola Music,
(strangely, to do with palettes...)
However, CarryMsgs are working, and moment/noteOffs,
moment/patch and moment/noteOns are being written
correctly in Study 1.
MidiChordSliderDefs (the moment/envs block) have not
begun to be programmed.
MidiRestDef has not been tested (but is probably okay).