Prevented the unnecessary repeated setting of controls.
Took advantage of the CarryMsgs to keep the state of
the controls that Moritz uses. If the state is the same as a
proposed message, the message is not saved to the file.
Shunting has to be done anyway in the Assistant Performer.
Changed the name of the inputChord and inputRest
attributes from "alignmentX" to "alignment".