Metrics restructuring #1
The (only) Metrics constructor now takes a CSSClass
argument, which is effectively the Metrics' _name_.
This means that all WriteSvg(...) functions will be
able to use (write) the known CSSClass, rather than
having to find it _again_.
This restructuring means changing most of the
Metrics subclasses, so it needs to be done in more
than one commit.
Also in the current commit: Used ClefIDs and FlagIDs
are now recorded in the ClefMetrics and
FlagsBockMetrics constructors.