Restructured PaletteGrps classes.
These now have a common PaletteGrps base class.
The PaletteGrps class has a ComposedBarlinePositions
function, which will be used for creating barlines
automatically (they are going to be overridable, of course).
This commit does not run. The construction of
barlines needs to be sorted out later on in Tombeau1.cs.