Major restructuring of the Tombeau1 code
Renamed the four prospective voices voice1 to
voice4. Their base class is now Tombeau1Voice.
Helper classes added: MsValues and Mode.
Added ModeSegment class, which is a list of
Grp objects (in the same voice), all of which
share the same Mode.
 A Mode is an immutable class, contaning a
list of absolute pitches (C=0, C#=1, D=2 etc.)
in order of importance, whereby not all
absolute pitches need to be included.
Modes can used, for example, to determine
the loudness or duration of particular pitches.