Added checks for invisible systems and tracks.
If a score contains a system that has no chords (and
is therefore completely invisible) a Debug.Assertion
fails.
If a score contains a track that has no chords (and
is therefore never visible on any system) a message
is displayed, and the score is created anyway.
Also:
Added private Generic ListToString<T>(...) function
to Moritz.Statics. Used it in a couple of special cases.
Added Voice.ContainsAChordSymbol property.