Check to enforce MidiChannel order (top to bottom of system).
Each track index (top to bottom) must now be the same as its MidiChannel. In other words, the top displayed track always has
MidiChannel == 0, and channels increase top to bottom of each
system. (error is fatal)
TODO: reprogram the Study 3 scores, so that this is the case. It
should still be possible to change the top to bottom display
order of the composed tracks, but the top displayed track
should always have MidiChannel==0.