Added the option to define performing regions in scores.
Regions are sections of a score that are repeated by
the Assistant Performer during performances.
The option has been implemented by defining a
non-abstract virtual SetScoreData(bars) function in
the CompositionAlgorithm class. This function
returns a null ScoreData object by default,
but can be overridden by any algorithm.
Null ScoreData objects are not written to scores.
Tested (and working) in the Tombeau1Algorithm.
If there is no score:scoreData object in a score, it
simply performs from beginning to end by default.