Study 4: Completed first test score version...
This score has seven pitchVector Trks and a Mode Trk containing chords having all the pitches in the above Trks. The Mode Trk is really just for ease of reading the current Mode. It might be better expressed as a series of arpeggios... Maybe implement this...
Bug?: Each pitchVector should, I think, be proceeding by minimum interval steps. Check this.