Study 4: Added UInt7 and PitchWeight structs...
This commit compiles, but does not run to completion of a Study 4 score.

UInt7 is defined in a new file in the Moritz.Globals assembly. It represents an integer in range 0..127, and is destined to percolate further into other classes such as MidiChordDef.
PitchWeight is defined (as two UInts) in a new file in Moritz.Spec.
Gamut is currently defined as a List of PitchWeight objects inside Mode. I think Gamut should really be a class in its own right, that replaces the Mode class.
TODO:
1. redefine Mode (and its contained Gamut)
2. correct ModeVector.Concat(ModeVector)