Trajectory path can now any number of nodes that is in the densityInputKrystal's shape...
The density input krystal determines the shape and levels of all the strands in the output. The trajectory node index is updated to match the shape of the density input krystal. The number of nodes in the  trajectory path must match the number of stands at some level in the density shape.