Removed the Study3Sketch2Algorithm and all the "Input..." classes it required
Input staves, including the InputStaff class and all the classes it used have been deleted.
The solution has been refactored accordingly, and now compiles and runs correctly.
All scores have been recompiled successfully, and seem to be in order.

This commit simplifies and reduces the codebase considerably, in preparation for more simplifications and further developments.