Revised Control handling in MidiChordDefs...
This is a temporary commit, that does not yet compile.
Each MidiChordDef contains a BasicMidiChordDefs and a MidiChordSliderDefs object.
These are distributed independently across the duration of the MidiChordDef.
BasicMidiChordDefs contain ResidentSynth controls that can be sent once per BasicMidiChordDef.
MidiChordSliderDefs contain ResidentSynth controls that can be sent independently (as sliders) over the course of the MidiChordDef.