Finalized Code:...
1. Removed redundant comments and code
2. Recompiled all the test scores
3. Checked the interpretations in the scores of Pianola Music and Study 1. They are as expected.
Pianola Music is automatically played 4 times, with differng interpretations (e.g. with organ and 1 random pitchWheel setting per chord).
Study 1 has two interpretation levels. The second interpretation (which will have to be selected by the user of the AssistantPerformer) is slower, and contains a random pitchWheel envelope per chord, whereby the pitchWheelSensitivity increases from 2 to 6 semitones as the interpretation progresses.