Pianola Music on 3 staves is now created correctly and is running in the AP....
Added a folder for the new score in the Moritz 2025 folder on OneDrive.
Copied the original .mkss file there (from the Pianola Music (version 2025) folder.
Edited the file in Moritz, so that there are 2 voices (midi channels) per staff.
The algorithm for the 3-staff version is identical to the original 6-staff version.